Fix loading bug
